+++ examples/create-react-app/src/index.js
@@ -1,8 +1,7 @@
 import * as React from 'react';
 import ReactDOM from 'react-dom';
-import StyledEngineProvider from '@material-ui/core/StyledEngineProvider';
 import CssBaseline from '@material-ui/core/CssBaseline';
-import { ThemeProvider } from '@material-ui/core/styles';
+import { ThemeProvider, StyledEngineProvider } from '@material-ui/core/styles';
 import App from './App';
 import theme from './theme';
