--- packages/material-ui/src/StyledEngineProvider/index.d.ts
@@ -1,1 +1,0 @@
-export { StyledEngineProvider as default } from '@material-ui/styled-engine';
