+++ packages/material-ui/src/index.d.ts
@@ -463,7 +463,6 @@ export { default as GlobalStyles } from './GlobalStyles';
 export * from './GlobalStyles';
 
 /**
- * @deprecated will be removed in v5.beta, please use StyledEngineProvider instead
+ * @deprecated will be removed in v5 beta, please use StyledEngineProvider instead
  */
-export { default as StylesProvider } from './StyledEngineProvider';
-export { default as StyledEngineProvider } from './StyledEngineProvider';
+export { StyledEngineProvider } from './styles';
