+++ packages/material-ui/src/index.js
@@ -401,5 +401,4 @@ export * from './GlobalStyles';
 /**
  * @deprecated will be removed in v5 beta, please use StyledEngineProvider instead
  */
-export { default as StylesProvider } from './StyledEngineProvider';
-export { default as StyledEngineProvider } from './StyledEngineProvider';
+export { StyledEngineProvider } from './styles';
