+++ packages/material-ui/src/styles/index.js
@@ -12,3 +12,4 @@ export { default as unstable_useThemeProps } from './useThemeProps';
 export { default as withStyles } from './withStyles';
 export { default as experimentalStyled } from './experimentalStyled';
 export { default as ThemeProvider } from './ThemeProvider';
+export { StyledEngineProvider } from '@material-ui/styled-engine';
