+++ test/regressions/TestViewer.js
@@ -1,8 +1,7 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import { useFakeTimers } from 'sinon';
-import StyledEngineProvider from '@material-ui/core/StyledEngineProvider';
-import { withStyles } from '@material-ui/core/styles';
+import { withStyles, StyledEngineProvider } from '@material-ui/core/styles';
 
 const styles = (theme) => ({
   '@global': {
